KaggleDataProcess/requirements.txt
altgraph==0.17
asgiref==3.3.1
cycler==0.10.0
Django==3.1.7
future==0.18.2
kiwisolver==1.3.1
numpy==1.20.1
pefile==2019.4.18
Pillow==8.1.2
psutil==5.8.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32-ctypes==0.2.0
six==1.15.0
sqlparse==0.4.1
